requirements.txt
asgiref==3.7.2
Django==4.2.3
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.6
Pygments==2.15.1
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
